Replaced many cases of enum with class enum. Fixed pragma once bug in bennyhill. Fixed file corruption generated by bennyhill. Added some error detection to ReadInput and WriteOutput, however static asserts triggered undefined compiler error in VS C++ ;-<

+++ include/rococo.api.h
@@ -229,7 +229,7 @@ namespace Rococo
 
 	template<typename T, typename U> [[nodiscard]] inline bool HasFlag(T flag, U flags)
 	{
-		return (flags & flag) != 0;
+		return ((int) flags & (int) flag) != 0;
 	}
 }
